Username and profile pic in top right
Got username and profile picture in login.js when we get the user role. I stored those in localstorage as well, and made a new script called "TopRightProfile" to actually display those elements from localstorage on the screen. Added that script to app.js to make sure it's used. Also updated usercontroller to pull the user's profile picture from the database as well.