QOL for dashboard
Added some buttons to navigate to the dashboard in the navbar. It pulls the user role from local storage to check whether to navigate to the normal dashboard, or the admin dashboard.

Also added a variable to check if a user is logged in. If true, it renders the buttons in the nav bar to go to the dashboard and COA. This doesn't appear to be working currently.