First attempt at CSS
I basically copied the JournalEntries.css script for CurrentUserCOA.css and attempted to add some css for the calendar, page container, buttons, due dates, tables, and email sections. This is the first css script I've made so it might be rough, but it seems like the styles match pretty well.